requirements.txt
requests==2.18.4
schematics==2.0.1
python-dotenv==0.7.1
Faker==0.8.11
pika==0.11.2